+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Added more buttons to support wdio tests for other themes (ex:mpages theme)
+
 6.2.0 - (June 12, 2019)
 ------------------
 ### Removed