+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -3,11 +3,14 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Added more buttons to support wdio tests for other themes (ex:mpages theme)
 
 6.2.0 - (June 12, 2019)
 ------------------
 ### Removed
 * Removed node_modules from .npmignore
+
 
 ### Changed
 * Replaced Terra-Form with Terra-Form-Checkbox in CollapsibleMenuViewToggle